pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "so101-control"
version = "0.1.0"
description = "Minimal SO101 robot arm control scripts"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5,<4.0",
    "feetech-servo-sdk>=1.0.0,<2.0.0",
]

[project.scripts]
so101-find-port = "find_port:main"
so101-setup-motors = "setup_motors:main"
so101-calibrate = "calibrate:main"
so101-control-panel = "control_panel:main"
